Rename jira package to utils
