+++ README.md
@@ -7,7 +7,7 @@ This repository provides a simple AI assistant that interacts with Jira.
 The backend is organized as a Python package under `backend/`. Key components include:
 
 - `actions/` – each file implements a single Sidekick action
-- `jira/` – helpers for interacting with Jira
+- `utils/` – helper functions for connecting to Jira and other utilities
 - `rag/` – scripts to build and query the FAISS index of Jira issues
 - `llm/` – utilities for language-model interactions
 - `context/` – context persistence
