+++ backend/actions/update_ticket.py
@@ -1,4 +1,4 @@
-from ..jira.jira_utils import update_ticket
+from ..utils.jira_utils import update_ticket
 
 
 def run(issue_key: str, comment: str):
