+++ backend/utils/get_info_from_jira.py
